refractored code with logic intact
