--- src/constants.js
@@ -1,13 +1,0 @@
-module.exports = {
-    BASE_URL: 'https://t<tenantId>.sandbox.mozu.com/api', // replace with your tenant id of your kibo sandbox 
-    ORDERS_URI: '/commerce/orders',
-    SHIPMENTS_URI: '/commerce/shipments',
-    ORDER_CANCEL_REASONS_URI: '/commerce/orders/cancel/reasons',
-    ORDER_CANCEL_URI: '/commerce/orders/cancel',
-    SET_FULFILLMENTINFO_URI: '/fulfillmentinfo',
-    OAUTH_URI: '/platform/applications/authtickets/oauth',
-    KIBO_CLIENT_ID: '<Client_Id>', // replace with your application client id created in your kibo sandbox
-    KIBO_CLIENT_SECRET: '<Client_Secret>' // replace with your client secret for your application created at kibo sandbox
-
-  };
-
