+++ src/helpers/constants.js
@@ -1,0 +1,12 @@
+module.exports = {
+    BASE_URL: 'https://t42183.sandbox.mozu.com/api', // replace with your tenant id of your kibo sandbox 
+    ORDERS_URI: '/commerce/orders',
+    SHIPMENTS_URI: '/commerce/shipments',
+    ORDER_CANCEL_REASONS_URI: '/commerce/orders/cancel/reasons',
+    ORDER_CANCEL_URI: '/commerce/orders/cancel',
+    SET_FULFILLMENTINFO_URI: '/fulfillmentinfo',
+    OAUTH_URI: '/platform/applications/authtickets/oauth',
+    KIBO_CLIENT_ID: 'irespons.KiboOMS_Sandbox.1.0.0.Release', // replace with your application client id created in your kibo sandbox
+    KIBO_CLIENT_SECRET: '68b0daa4b7d84308b2f47c0822c84be8' // replace with your client secret for your application created at kibo sandbox
+};
+
